TST: move zip_latest tests
