+++ shed/tests/test_streams.py
@@ -14,7 +14,6 @@
 ##############################################################################
 # TODO: this should be broken up into a module per node, that way we
 # can test individual nodes without having to test everything
-from itertools import zip_longest
 
 import pytest
 from numpy.testing import assert_equal
@@ -86,229 +85,6 @@ def test_combine_latest(exp_db, start_uid1, start_uid3):
         assert l1 != l2
         if name == 'event':
             assert l2['seq_num'] == 2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-
-def test_zip_latest(exp_db, start_uid1, start_uid3):
-    source = Stream()
-    source2 = Stream()
-
-    L = es.zip_latest(source, source2).sink_to_list()
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s = exp_db.restream(ih1)
-    s2 = exp_db.restream(ih2)
-    for b in s2:
-        source2.emit(b)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for name, (l1, l2) in L:
-        assert_docs.add(name)
-        assert l1 != l2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-    assert len(L) == len(list(exp_db.restream(ih1)))
-
-
-def test_zip_latest_reverse(exp_db, start_uid1, start_uid3):
-    source = Stream()
-    source2 = Stream()
-
-    L = es.zip_latest(source, source2).sink_to_list()
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s = exp_db.restream(ih1)
-    s2 = exp_db.restream(ih2)
-    for a in s:
-        source.emit(a)
-    for b in s2:
-        source2.emit(b)
-
-    assert_docs = set()
-    for name, (l1, l2) in L:
-        assert_docs.add(name)
-        assert l1 != l2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-    assert len(L) == len(list(exp_db.restream(ih1)))
-
-
-def test_zip_latest_double(exp_db, start_uid1, start_uid3):
-    source = Stream()
-    source2 = Stream()
-
-    dp = es.zip_latest(source, source2)
-    L = dp.sink_to_list()
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s = exp_db.restream(ih1)
-    s2 = exp_db.restream(ih2)
-    for b in s2:
-        source2.emit(b)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for name, (l1, l2) in L:
-        assert_docs.add(name)
-        assert l1 != l2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-    assert len(L) == len(list(exp_db.restream(ih1)))
-
-    L.clear()
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s = exp_db.restream(ih1)
-    s2 = exp_db.restream(ih2)
-    for b in s2:
-        source2.emit(b)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for name, (l1, l2) in L:
-        print(l1)
-        assert_docs.add(name)
-        assert l1 != l2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-    assert len(L) == len(list(exp_db.restream(ih1)))
-
-
-def test_zip_latest_double_reverse(exp_db, start_uid1, start_uid3):
-    source = Stream()
-    source2 = Stream()
-
-    L = es.zip_latest(source, source2).sink_to_list()
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s = exp_db.restream(ih1)
-    s2 = exp_db.restream(ih2)
-    for a in s:
-        source.emit(a)
-    for b in s2:
-        source2.emit(b)
-
-    assert_docs = set()
-    for name, (l1, l2) in L:
-        assert_docs.add(name)
-        assert l1 != l2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-    assert len(L) == len(list(exp_db.restream(ih1)))
-
-    L.clear()
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s = exp_db.restream(ih1)
-    s2 = exp_db.restream(ih2)
-    for a in s:
-        source.emit(a)
-    for b in s2:
-        source2.emit(b)
-
-    assert_docs = set()
-    for name, (l1, l2) in L:
-        assert_docs.add(name)
-        assert l1 != l2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-    assert len(L) == len(list(exp_db.restream(ih1)))
-
-
-def test_zip_latest_double_interleaved(exp_db, start_uid1, start_uid3):
-    source = Stream()
-    source2 = Stream()
-
-    dp = es.zip_latest(source, source2)
-    L = dp.sink_to_list()
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s = exp_db.restream(ih1)
-    s2 = exp_db.restream(ih2)
-    for b in s2:
-        source2.emit(b)
-    for a in s:
-        source.emit(a)
-
-    assert_docs = set()
-    for name, (l1, l2) in L:
-        assert_docs.add(name)
-        assert l1 != l2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-    assert len(L) == len(list(exp_db.restream(ih1)))
-
-    L.clear()
-    ih1 = exp_db[start_uid1]
-    ih2 = exp_db[start_uid3]
-    s = exp_db.restream(ih1)
-    s2 = exp_db.restream(ih2)
-    for b, a in zip_longest(s2, s):
-        if a:
-            source.emit(a)
-        if b:
-            source2.emit(b)
-
-    assert_docs = set()
-    for name, (l1, l2) in L:
-        assert_docs.add(name)
-        assert l1 != l2
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-    assert len(L) == len(list(exp_db.restream(ih1)))
-
-
-def test_zip_latest_double_clear():
-    source = Stream()
-    source2 = Stream()
-
-    dp = es.zip_latest(source, source2, clear_on_lossless_stop=True)
-    L = dp.sink_to_list()
-    g1 = list(to_event_model([1, 2, 3], [('det', {})]))
-    g2 = list(to_event_model(['a', 'b', 'c'], [('det', {})]))
-    g3 = list(to_event_model([5], [('det', {})]))
-    for b in g1:
-        source.emit(b)
-    for a in g2:
-        source.emit(a)
-    for c in g3:
-        source2.emit(c)
-
-    assert_docs = set()
-    assert len(L) == len(g2)
-    for (name, (l1, l2)), (_, ll1) in zip(L, g2):
-        assert_docs.add(name)
-        assert l1 != l2
-        assert l1 == ll1
-        if name == 'event':
-            assert l2['data']['det'] == 5
-    for n in ['start', 'descriptor', 'event', 'stop']:
-        assert n in assert_docs
-
-    L.clear()
-    g1 = list(to_event_model([10, 20, 30], [('det', {})]))
-    g2 = list(to_event_model(['aa', 'bb', 'cc'], [('det', {})]))
-    g3 = list(to_event_model([50], [('det', {})]))
-    for b in g1:
-        source.emit(b)
-    for a in g2:
-        source.emit(a)
-    for c in g3:
-        source2.emit(c)
-
-    assert_docs = set()
-    assert len(L) == len(g2)
-    for (name, (l1, l2)), (_, ll1) in zip(L, g2):
-        assert_docs.add(name)
-        assert l1 != l2
-        assert l1 == ll1
-        if name == 'event':
-            assert l2['data']['det'] == 50
     for n in ['start', 'descriptor', 'event', 'stop']:
         assert n in assert_docs
